Add start and fix commands to index
